+++ tests/utils/deployMockSystem.js
@@ -159,9 +159,7 @@ const deployMockSystem = async (
       shares.options.address,
       vault.options.address,
       registry.options.address,
-      fundFactory.options.address,
-      engine.options.address,
-      mln.options.address,
+      fundFactory.options.address
     ])
     .send(defaultTxOpts);